Remove Log::Any::Adapter::Term dependency.
Moved Log::Any::Adapter::Term into this package as NuFoo::LogAnyAdapter. The
old plan was to release the adapter as a separate module but after discussions
with the author (Jonathan Swartz) it was decided not to, as it doesn't fit the
Log::Any idea to bridge to real logging frameworks (rather than squeeze one
into the adapter as I did).

There is a good chance that NuFoo will just move to Log::Dispatch at some point
as there is a mod there that does color term logging plus we can then add
logging to a file as well.